fix: asdf installation more robust

+++ all-install.sh
@@ -118,5 +118,5 @@ fish -c "fisher update"
 # https://asdf-vm.com/guide/getting-started.html
 # dotfiles/fish/conf.d/asdf.fish has the Fish shell related iniatlization
 [ -d ~/.asdf ] || git clone https://github.com/asdf-vm/asdf.git ~/.asdf --branch v0.13.1
-ln -s ~/.asdf/completions/asdf.fish ~/.config/fish/completions
-asdf plugin-add deno https://github.com/asdf-community/asdf-deno.git
+[ -s ~/.config/fish/completions/asdf.fish ] || ln -s ~/.asdf/completions/asdf.fish ~/.config/fish/completions
+~/.asdf/bin/asdf plugin-add deno https://github.com/asdf-community/asdf-deno.git
